[BINS]: mbedtls only linked in the kernel

+++ build/autobuild.sh
@@ -42,6 +42,9 @@ aos make clean > /dev/null 2>&1
 for target in ${mk3060_targets}; do
     for platform in ${mk3060_platforms}; do
 	for bins in ${bins_type}; do
+            if [ ${target} = "tls" ]; then
+                continue
+            fi
             aos make ${target}@${platform} BINS=${bins} > ${target}@${platform}@${bins}@${branch}.multi-bins.log 2>&1
             if [ -f out/${target}@${platform}/binary/${target}@${platform}.${bins}.elf ]; then
                 rm -rf ${target}@${platform}@${bins}@${branch}.multi-bins.log
